refactor : firebase integration and supabase application
