+++ .env
@@ -1,10 +1,14 @@
-API_TOKEN=
-DB_PASSWORD=hd121612
-DB_URL=jdbc:postgresql://cpdb.cf6k82ssc832.ap-northeast-2.rds.amazonaws.com:5432/cpdb
-DB_USERNAME=postgres
-JWT_SECRET=J25P6nDtgArp8fljXAwPiiQaQvc6N0or
-R2_ACCESS_KEY=5abfeb699aefbae399603640cbd75ff3
-R2_BUCKET=cp-main-be
-R2_ENDPOINT=https://a35471a7db408765c522b98a7b65d786.r2.cloudflarestorage.com/cp-main-be
-R2_SECRET_KEY=72834ca03a49383c5893e3d567ecb1a17de78ab1c21dd73e871ce8a11216bfd1
-FASTAPI_URL=http://13.54.125.183:8000
+API_TOKEN=;
+DB_PASSWORD=Camping123!@##;
+
+DB_URL=jdbc:postgresql://aws-1-ap-south-1.pooler.supabase.com:5432/postgres?sslmode=require;
+DB_USERNAME=postgres.zgjxxldhyhofpmrwdsbh;
+FASTAPI_URL=http://43.200.84.255:8000;
+JWT_SECRET=gJlHIv7uh8RmCenueiYqdZPKHVwoc2ZpJqJCwImLceI=;
+R2_ACCESS_KEY=5abfeb699aefbae399603640cbd75ff3;
+R2_ACCOUNT_ID=a35471a7db408765c522b98a7b65d786;
+R2_BUCKET=cp-main-be;
+R2_ENDPOINT=https://a35471a7db408765c522b98a7b65d786.r2.cloudflarestorage.com;
+R2_SECRET_KEY=72834ca03a49383c5893e3d567ecb1a17de78ab1c21dd73e871ce8a11216bfd1;
+SUPABASE_SERVICE_KEY=sb_secret_ryQLXVV0VArQMen2cG0sfQ_7jifMfi5;
+SUPABASE_URL=https://zgjxxldhyhofpmrwdsbh.supabase.co
