+++ .gitignore
@@ -39,3 +39,4 @@ out/
 
 # Firebase
 src/main/resources/napulnapul-d9572-firebase-adminsdk-fbsvc-0a2757f074.json
+
