+++ build.gradle
@@ -64,6 +64,7 @@ dependencies {
 	testImplementation 'org.testcontainers:junit-jupiter'
 	testImplementation 'com.squareup.okhttp3:mockwebserver:4.12.0'
 	testImplementation 'com.squareup.okio:okio:3.9.0'
+	runtimeOnly 'org.postgresql:postgresql'
 }
 
 tasks.named('test') {
